requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
discord.py==1.7.3
frozenlist==1.3.0
idna==3.3
multidict==5.2.0
requests==2.27.1
typing_extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
